requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-firestore
pandas
python-dotenv
requests
